requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-WTF
Flask-Login
Flask-Migrate
Werkzeug
